+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## Unreleased
+## 0.4.0 (Unreleased)
   * clean up FunctorRec into standalone HFunctorList module
+  * clean up HexBytestring into standalone module
 
 ## 0.3.1 (2021-12-21)
   * packaging improvements (documentation, CI)